requirements.txt
telethon==1.36.0
fastapi==0.115.5
uvicorn==0.32.0
python-dotenv==1.0.1
qrcode